bdev/ec: widen the WIB and unmapped-bitmap generation counters to u64
longhorn/longhorn-1061

+++ module/bdev/ec/bdev_ec.h
@@ -222,7 +222,7 @@ int ec_bdev_stop_rebuild(const char *ec_name);
 int ec_bdev_get_wib_status(const char *ec_name,
 			   uint32_t   *num_regions,
 			   uint32_t   *dirty_regions,
-			   uint32_t   *generation,
+			   uint64_t   *generation,
 			   bool       *persist_pending);
 
 /*
@@ -247,7 +247,7 @@ int ec_bdev_get_unmap_status(const char *ec_name,
 			     uint64_t   *num_stripes,
 			     uint64_t   *unmapped_stripes,
 			     uint64_t   *blob_bytes,
-			     uint32_t   *generation,
+			     uint64_t   *generation,
 			     uint8_t    *active_copy,
 			     bool       *persist_pending);
